+++ frontend/src/App.js
@@ -2,7 +2,7 @@ import { Navigate, Route, Routes } from 'react-router-dom';
 import './App.css';
 import Login from './pages/login/login';
 import Signup from './pages/signup/signup';
-import Home from './pages/helppp/home'
+import Home from './pages/home/home'
 
 
 function App() {
